requirements.txt
bleach==2.0.0
certifi==2017.4.17
chardet==3.0.4
Django==1.11.3
django-simple-history==1.9.0
django-widget-tweaks==1.4.1
html5lib==0.999999999
idna==2.5
Markdown==2.6.8
pkg-resources==0.0.0
postmarker==0.11.2
psycopg2==2.7.1
pytz==2017.2
requests==2.18.1
six==1.10.0
urllib3==1.21.1
webencodings==0.5.1
